# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PharmaCart synthetic API",
    "version": "0.2.0",
    "description": "Local B0-B3 backend. Cash EGP tax-exempt fixtures only; no real suppliers."
  },
  "security": [
    {
      "oidc": []
    }
  ],
  "components": {
    "securitySchemes": {
      "oidc": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "correlationId"
            ],
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1,
                "maxLength": 64,
                "pattern": "^[A-Za-z][A-Za-z0-9_]*$",
                "description": "Extensible refusal code. Boundary and tenant codes are upper case (INVALID_REQUEST, IDEMPOTENCY_KEY_SCOPE_CONFLICT); an inventory refusal forwards the domain rejection reason verbatim in lower case (stale_sequence, conflicting_partition), so the code is deliberately not enumerated."
              },
              "message": {
                "type": "string",
                "minLength": 1,
                "description": "Fixed redacted sentence. It never echoes request values or internal detail."
              },
              "correlationId": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        }
      },
...
      "TenantContext": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "principalKind",
          "userSubject",
          "membershipId",
          "organisationId",
          "organisationKind",
          "branchId",
          "allowedBranchIds",
          "role",
          "membershipVersion"
        ],
        "description": "Scope resolved for this request from the verified token subject and the requested organisation and branch headers. It carries no token, no credential and no other membership.",
        "properties": {
          "principalKind": {
            "const": "member",
            "description": "Only member principals reach the tenant routes; service principals are not implemented."
          },
          "userSubject": {
            "type": "string",
            "minLength": 1,
            "description": "Verified access token subject."
          },
          "membershipId": {
            "type": "string",
            "format": "uuid"
          },
          "organisationId": {
            "type": "string",
            "format": "uuid"
          },
          "organisationKind": {
            "enum": [
              "pharmacy",
              "supplier"
            ],
            "description": "Purchasing and receipt routes additionally require a pharmacy organisation."
          },
          "branchId": {
            "type": "string",
            "format": "uuid"
          },
          "allowedBranchIds": {
            "type": "array",
            "minItems": 1,
            "maxItems": 1,
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Exactly the requested branch: one transaction never widens beyond the branch its row level security policy is set to."
          },
          "role": {
            "enum": [
              "pharmacy_owner",
              "purchaser",
              "receiver",
              "supplier_operator",
              "supplier_administrator",
              "support"
            ],
            "description": "Active membership role. Quoting and approval require pharmacy_owner or purchaser; receipts require pharmacy_owner or receiver. Need and order lists and mapping candidates require pharmacy_owner or purchaser; binding a mapping requires pharmacy_owner."
          },
          "membershipVersion": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991,
            "description": "Version of the active membership row used for this request."
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/context": {
      "get": {
        "operationId": "getContext",
        "parameters": [
          {
            "name": "X-Organisation-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Branch-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resolved membership scope for the request",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantContext"
                }
              }
            }
          },
          "400": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/needs/{id}": {
      "get": {
        "operationId": "getNeed",
        "parameters": [
          {
            "name": "X-Organisation-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Branch-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The requested need inside the resolved scope",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeedView"
                }
              }
            }
          },
          "400": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/inventory": {
      "post": {
        "operationId": "ingestInventory",
        "description": "Installation scope derives from the verified subject. Durable processing completes before acknowledgement.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "additionalProperties": false,
                    "required": [
                      "kind",
                      "eventId",
                      "snapshotId",
                      "sequence",
                      "partitionId",
                      "rows"
                    ],
                    "properties": {
                      "eventId": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 256,
                        "pattern": "^[^\\u0000]+$"
                      },
                      "snapshotId": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 256,
                        "pattern": "^[^\\u0000]+$"
                      },
                      "sequence": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 9007199254740991
                      },
                      "kind": {
                        "const": "partition"
                      },
                      "partitionId": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 256,
                        "pattern": "^[^\\u0000]+$"
                      },
                      "rows": {
                        "type": "array",
                        "maxItems": 1000,
                        "items": {
                          "type": "object",
                          "additionalProperties": false,
                          "required": [
                            "sourceCode",
                            "quantity",
                            "unit"
                          ],
                          "properties": {
                            "sourceCode": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 256,
                              "pattern": "^[^\\u0000]+$"
                            },
                            "unit": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 256,
                              "pattern": "^[^\\u0000]+$"
                            },
                            "quantity": {
                              "type": "string",
                              "maxLength": 128,
                              "pattern": "^(?:0|0\\.[0-9]*[1-9]|[1-9][0-9]*(?:\\.[0-9]*[1-9])?)$"
                            }
                          }
                        }
                      }
                    }
                  },
                  {
                    "type": "object",
                    "additionalProperties": false,
                    "required": [
                      "kind",
                      "eventId",
                      "snapshotId",
                      "sequence",
                      "expectedPartitionIds"
                    ],
                    "properties": {
                      "eventId": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 256,
                        "pattern": "^[^\\u0000]+$"
                      },
                      "snapshotId": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 256,
                        "pattern": "^[^\\u0000]+$"
                      },
                      "sequence": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 9007199254740991
                      },
                      "kind": {
                        "const": "complete"
                      },
                      "expectedPartitionIds": {
                        "type": "array",
                        "minItems": 1,
                        "maxItems": 1000,
                        "uniqueItems": true,
                        "items": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 256,
                          "pattern": "^[^\\u0000]+$"
                        }
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Snapshot event processed durably",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryAcceptance"
                }
              }
            }
          },
          "400": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/orders/{id}": {
      "get": {
        "operationId": "getOrder",
        "parameters": [
          {
            "name": "X-Organisation-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Branch-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The requested order intent inside the resolved scope",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            }
          },
          "400": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Redacted error",
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  }
}